bidding_interface/bidding/requirements.txt
flask==3.1.0
numpy==1.26.4
pandas==2.2.3
tensorflow-cpu==2.16.1
scikit-learn==1.6.1
werkzeug==3.1.3
jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
colorama==0.4.6
markupsafe==3.0.2
h5py>=3.11.0
ml-dtypes~=0.3.1 